Added validation and Connection Exception

+++ src/main/java/ca/paymentrails/paymentrails/PaymentRails_Batch.java
@@ -1,5 +1,7 @@
 package ca.paymentrails.paymentrails;
 
+import ca.paymentrails.Exceptions.InvalidConnectionException;
+import ca.paymentrails.Exceptions.InvalidFieldException;
 import ca.paymentrails.Exceptions.InvalidStatusCodeException;
 
 /**
@@ -13,9 +15,15 @@ public class PaymentRails_Batch {
      *
      * @param batch_id
      * @return The response
-     * @throws InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidFieldException
+     * @throws ca.paymentrails.Exceptions.InvalidConnectionException
      */
-    public static String get(String batch_id) throws InvalidStatusCodeException {
+    public static String get(String batch_id) throws InvalidStatusCodeException, InvalidFieldException, InvalidConnectionException {
+        if(batch_id == null || batch_id.isEmpty()){
+            throw new InvalidFieldException("Batch id cannot be null or empty.");
+        }
+
         PaymentRails_Client client = PaymentRails_Client.create();
 
         String endPoint = "v1/batches/" + batch_id;
@@ -29,9 +37,17 @@ public class PaymentRails_Batch {
      * @param batch_id
      * @param body
      * @return The response
-     * @throws InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidFieldException
+     * @throws ca.paymentrails.Exceptions.InvalidConnectionException
      */
-    public static String patch(String batch_id, String body) throws InvalidStatusCodeException {
+    public static String patch(String batch_id, String body) throws InvalidStatusCodeException, InvalidFieldException, InvalidConnectionException {
+         if(batch_id == null || batch_id.isEmpty()){
+            throw new InvalidFieldException("Batch id cannot be null or empty.");
+        }
+        if(body == null || body.isEmpty()){
+            throw new InvalidFieldException("Body cannot be null or empty.");
+        }
         PaymentRails_Client client = PaymentRails_Client.create();
         String endPoint = "v1/batches/" + batch_id;
         String response = client.patch(endPoint, body);
@@ -43,9 +59,15 @@ public class PaymentRails_Batch {
      *
      * @param batch_id
      * @return The response
-     * @throws InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidFieldException
+     * @throws ca.paymentrails.Exceptions.InvalidConnectionException
      */
-    public static String delete(String batch_id) throws InvalidStatusCodeException {
+    public static String delete(String batch_id) throws InvalidStatusCodeException, InvalidFieldException, InvalidConnectionException {
+         if(batch_id == null || batch_id.isEmpty()){
+            throw new InvalidFieldException("Batch id cannot be null or empty.");
+        }
+
         PaymentRails_Client client = PaymentRails_Client.create();
         String endPoint = "v1/batches/" + batch_id;
         String response = client.delete(endPoint);
@@ -57,9 +79,14 @@ public class PaymentRails_Batch {
      *
      * @param body
      * @return The response
-     * @throws InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidFieldException
+     * @throws ca.paymentrails.Exceptions.InvalidConnectionException
      */
-    public static String post(String body) throws InvalidStatusCodeException {
+    public static String post(String body) throws InvalidStatusCodeException, InvalidFieldException, InvalidConnectionException {
+        if(body == null || body.isEmpty()){
+            throw new InvalidFieldException("Body cannot be null or empty.");
+        }
         PaymentRails_Client client = PaymentRails_Client.create();
         String endPoint = "v1/batches/";
         String response = client.post(endPoint, body);
@@ -71,9 +98,15 @@ public class PaymentRails_Batch {
      *
      * @param batch_id
      * @return The response
-     * @throws InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidFieldException
+     * @throws ca.paymentrails.Exceptions.InvalidConnectionException
      */
-    public static String generateQuote(String batch_id) throws InvalidStatusCodeException {
+    public static String generateQuote(String batch_id) throws InvalidStatusCodeException, InvalidFieldException, InvalidConnectionException {
+        if(batch_id == null || batch_id.isEmpty()){
+            throw new InvalidFieldException("Batch id cannot be null or empty.");
+        }
+
         PaymentRails_Client client = PaymentRails_Client.create();
         String endPoint = "v1/batches/" + batch_id + "/generate-quote";
         String response = client.post(endPoint);
@@ -85,9 +118,14 @@ public class PaymentRails_Batch {
      *
      * @param batch_id
      * @return The response
-     * @throws InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidFieldException
+     * @throws ca.paymentrails.Exceptions.InvalidConnectionException
      */
-    public static String processBatch(String batch_id) throws InvalidStatusCodeException {
+    public static String processBatch(String batch_id) throws InvalidStatusCodeException, InvalidFieldException, InvalidConnectionException {
+        if(batch_id == null || batch_id.isEmpty()){
+            throw new InvalidFieldException("Batch id cannot be null or empty.");
+        }
         PaymentRails_Client client = PaymentRails_Client.create();
         String endPoint = "v1/batches/" + batch_id + "/start-processing";
         String response = client.post(endPoint);
@@ -102,9 +140,20 @@ public class PaymentRails_Batch {
      * @param pageSize
      * @param message
      * @return The response
-     * @throws InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidConnectionException
+     * @throws ca.paymentrails.Exceptions.InvalidFieldException
      */
-    public static String query(int page, int pageSize, String message) throws InvalidStatusCodeException {
+    public static String query(int page, int pageSize, String message) throws InvalidStatusCodeException, InvalidConnectionException, InvalidFieldException {
+        if (page < 0) {
+            throw new InvalidFieldException("Page cannot be less then 0");
+        }
+        if (pageSize < 0) {
+            throw new InvalidFieldException("Page size cannot be less then 0");
+        }
+        if (message == null) {
+            throw new InvalidFieldException("Message cannot be null");
+        }
         PaymentRails_Client client = PaymentRails_Client.create();
         String endPoint = "v1/recipients/?" + "&search=" + message + "&page=" + page + "&pageSize=" + pageSize;
         String response = client.get(endPoint);
@@ -116,9 +165,11 @@ public class PaymentRails_Batch {
      *
      * @param message
      * @return The response
-     * @throws InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidConnectionException
+     * @throws ca.paymentrails.Exceptions.InvalidFieldException
      */
-    public static String query(String message) throws InvalidStatusCodeException {
+    public static String query(String message) throws InvalidStatusCodeException, InvalidConnectionException, InvalidFieldException {
         return query(1, 10, message);
     }
 
@@ -126,9 +177,11 @@ public class PaymentRails_Batch {
      * List all batches
      *
      * @return The response
-     * @throws InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidConnectionException
+     * @throws ca.paymentrails.Exceptions.InvalidFieldException
      */
-    public static String query() throws InvalidStatusCodeException {
+    public static String query() throws InvalidStatusCodeException, InvalidConnectionException, InvalidFieldException {
         return query(1, 10, "");
     }
 
@@ -139,9 +192,11 @@ public class PaymentRails_Batch {
      * @param page
      * @param pageNumber
      * @return The response
-     * @throws InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidConnectionException
+     * @throws ca.paymentrails.Exceptions.InvalidFieldException
      */
-    public static String query(int page, int pageNumber) throws InvalidStatusCodeException {
+    public static String query(int page, int pageNumber) throws InvalidStatusCodeException, InvalidConnectionException, InvalidFieldException {
         return query(page, pageNumber, "");
     }
 
@@ -150,9 +205,14 @@ public class PaymentRails_Batch {
      *
      * @param batch_id
      * @return The response
-     * @throws InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidStatusCodeException
+     * @throws ca.paymentrails.Exceptions.InvalidFieldException
+     * @throws ca.paymentrails.Exceptions.InvalidConnectionException
      */
-    public static String summary(String batch_id) throws InvalidStatusCodeException {
+    public static String summary(String batch_id) throws InvalidStatusCodeException, InvalidFieldException, InvalidConnectionException {
+        if(batch_id == null || batch_id.isEmpty()){
+            throw new InvalidFieldException("Batch id cannot be null os empty");
+        }
         PaymentRails_Client client = PaymentRails_Client.create();
 
         String endPoint = "v1/batches/" + batch_id + "/summary";
